# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qi local laboratory",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/experiments": {
      "get": {
        "summary": "Experiments",
        "operationId": "experiments_api_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunEntry"
                  },
                  "type": "array",
                  "title": "Response Experiments Api Experiments Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{run_id}": {
      "get": {
        "summary": "Detail",
        "operationId": "detail_api_experiments__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportData"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/api/learn/generation": {
      "get": {
        "summary": "Lesson",
        "operationId": "lesson_api_learn_generation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationLesson"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/players": {
      "get": {
        "summary": "Players",
        "operationId": "players_api_players_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfo"
                  },
                  "type": "array",
                  "title": "Response Players Api Players Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/new": {
      "post": {
        "summary": "New",
        "operationId": "new_api_new_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/inspect": {
      "post": {
        "summary": "Position",
        "operationId": "position_api_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/apply": {
      "post": {
        "summary": "Apply",
        "operationId": "apply_api_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/play/controller/inspect": {
      "post": {
        "summary": "Controller",
        "operationId": "controller_api_play_controller_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Controller"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Controller"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/play/choose": {
      "post": {
        "summary": "Play",
        "operationId": "play_api_play_choose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/play/session/inspect": {
      "post": {
        "summary": "Session",
        "operationId": "session_api_play_session_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApplyRequest": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "move": {
            "type": "string",
            "maxLength": 4,
            "minLength": 4,
            "title": "Move"
          },
          "expected_state_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected State Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot",
          "move",
          "expected_state_hash"
        ],
        "title": "ApplyRequest"
      },
...
      "Controller": {
        "properties": {
          "player": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Player",
            "default": "human"
          },
          "settings": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "maxProperties": 8,
            "title": "Settings"
          },
          "binding_sha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding Sha256"
          },
          "checkpoint_sha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Controller",
        "description": "A UI selection stores identity, never a resource path."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "GenerationLesson": {
        "properties": {
          "example": {
            "$ref": "#/components/schemas/LessonExample"
          },
          "phase_policy": {
            "type": "string",
            "title": "Phase Policy"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/LessonFrame"
            },
            "type": "array",
            "title": "Frames"
          }
        },
        "type": "object",
        "required": [
          "example",
          "phase_policy",
          "frames"
        ],
        "title": "GenerationLesson"
      },
...
      "InspectRequest": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot"
        ],
        "title": "InspectRequest"
      },
...
      "PlayRequest": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "expected_state_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected State Hash"
          },
          "controller": {
            "$ref": "#/components/schemas/Controller"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot",
          "expected_state_hash",
          "controller"
        ],
        "title": "PlayRequest"
      },
      "PlayResult": {
        "properties": {
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "choice": {
            "$ref": "#/components/schemas/Choice"
          },
          "config": {
            "$ref": "#/components/schemas/PlayerConfig"
          }
        },
        "type": "object",
        "required": [
          "position",
          "choice",
          "config"
        ],
        "title": "PlayResult"
      },
...
      "PlayerInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "uses_search": {
            "type": "boolean",
            "title": "Uses Search"
          },
          "default_nodes": {
            "type": "integer",
            "title": "Default Nodes",
            "default": 128
          },
          "default_depth": {
            "type": "integer",
            "title": "Default Depth",
            "default": 2
          },
          "checkpoint_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Sha256"
          },
          "default_rollout_plies": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Rollout Plies"
          },
          "implementation_id": {
            "type": "string",
            "title": "Implementation Id",
            "default": ""
          },
          "binding_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding Sha256"
          },
          "available": {
            "type": "boolean",
            "title": "Available",
            "default": true
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "settings": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Setting"
            },
            "type": "object",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "label",
          "description",
          "uses_search"
        ],
        "title": "PlayerInfo"
      },
      "Position": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "board": {
            "type": "string",
            "title": "Board"
          },
          "turn": {
            "type": "string",
            "enum": [
              "red",
              "black"
            ],
            "title": "Turn"
          },
          "ply": {
            "type": "integer",
            "title": "Ply"
          },
          "state_hash": {
            "type": "string",
            "title": "State Hash"
          },
          "legal_moves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Legal Moves"
          },
          "in_check": {
            "type": "boolean",
            "title": "In Check"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Result"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "snapshot",
          "board",
          "turn",
          "ply",
          "state_hash",
          "legal_moves",
          "in_check",
          "outcome"
        ],
        "title": "Position"
      },
...
      "ReportData": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "kind": {
            "type": "string",
            "const": "search",
            "title": "Kind",
            "default": "search"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "planned": {
            "type": "integer",
            "title": "Planned"
          },
          "validation": {
            "type": "string",
            "title": "Validation"
          },
          "evidence_sha256": {
            "type": "string",
            "title": "Evidence Sha256"
          },
          "presentation_sha256": {
            "type": "string",
            "title": "Presentation Sha256"
          },
          "glossary": {
            "$ref": "#/components/schemas/Glossary"
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "by_position": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Summary"
            },
            "type": "object",
            "title": "By Position"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitSummary"
            },
            "type": "array",
            "title": "Units"
          },
          "traces": {
            "items": {
              "$ref": "#/components/schemas/TraceSummary"
            },
            "type": "array",
            "title": "Traces"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "narrative": {
            "type": "string",
            "title": "Narrative"
          },
          "narrative_sha256": {
            "type": "string",
            "title": "Narrative Sha256"
          },
          "trace_unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Unavailable Reason"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "provenance",
          "status",
          "completed",
          "planned",
          "validation",
          "evidence_sha256",
          "presentation_sha256",
          "glossary",
          "summary",
          "by_position",
          "units",
          "traces",
          "warnings",
          "narrative",
          "narrative_sha256",
          "trace_unavailable_reason"
        ],
        "title": "ReportData"
      },
...
      "RunEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "location",
          "kind",
          "status"
        ],
        "title": "RunEntry"
      },
...
      "SessionResult": {
        "properties": {
          "session": {
            "$ref": "#/components/schemas/GameSession"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          }
        },
        "type": "object",
        "required": [
          "session",
          "position"
        ],
        "title": "SessionResult"
      },
...
    }
  }
}